requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.1
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.7
gevent==24.2.1
greenlet==3.1.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
kombu==5.4.1
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.1
prompt_toolkit==3.0.47
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
setuptools==75.1.0
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
zope.event==5.0
zope.interface==7.0.3
